Fixed bug.
Fixed a bug where the settings where it was possible to change the
settings even though they were "hidden". The settings container was
layered on top of the main page, but was not visible due to the
transparency setting given to it; so although it was invisible, it was
possible to still interact with it. This is a quick fix that instead
sets the display of the settings to none and its default block to
display the settings card.